# api/assets/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "GK Circle PCS examination API",
    "version": "0.0.1-alpha"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/v1/admin/reports/list": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Reports"],
        "summary": "Get details of quizzes list host by Admin.",
        "operationId": "RequestListQuizzesAnalysis",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "OrderBy",
            "name": "orderBy",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/RsponseListQuizzesAnalysis"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/admin/reports/{active_quiz_id}/analysis": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Reports"],
        "summary": "Get details of quizzes host by Admin.",
        "operationId": "RequestGetQuizAnalysis",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActiveQuizId",
            "name": "active_quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseGetQuizAnalysis"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
    "/v1/quizzes": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Quiz"],
        "summary": "Get details of quizzes uploaded by Admin.",
        "operationId": "GetAdminUploadedQuizzes",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseAdminUploadedQuiz"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/quizzes/{quiz_id}": {
      "delete": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Quiz"],
        "summary": "Delete quiz that created by user (if no active quiz is present).",
        "operationId": "DeleteQuizById",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/quizzes/{quiz_id}/demo_session": {
      "post": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Quiz"],
        "summary": "Create quiz active for user.",
        "operationId": "RequestGenerateDemoSession",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseGenerateDemoSession"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
    "/v1/quizzes/{quiz_title}/upload": {
      "post": {
        "consumes": ["multipart/form-data"],
        "schemes": ["http", "https"],
        "tags": ["Quiz"],
        "summary": "Create a new quiz by uploading a CSV file.",
        "operationId": "RequestQuizCreated",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizTitle",
            "name": "quiz_title",
            "in": "path",
            "required": true
          },
          {
            "type": "file",
            "x-go-name": "File",
            "name": "attachment",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Description",
            "name": "description",
            "in": "formData"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseQuizCreated"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "NullString": {
      "description": "var s NullString\nerr := db.QueryRow(\"SELECT name FROM foo WHERE id=?\", id).Scan(\u0026s)\n...\nif s.Valid {\nuse s.String\n} else {\nNULL value\n}",
      "type": "object",
      "title": "NullString represents a string that may be null.\nNullString implements the Scanner interface so\nit can be used as a scan destination:",
      "properties": {
        "String": {
          "type": "string"
        },
        "Valid": {
          "type": "boolean"
        }
      },
      "x-go-package": "database/sql"
    },
    "NullTime": {
      "description": "NullTime implements the Scanner interface so\nit can be used as a scan destination, similar to NullString.",
      "type": "object",
      "title": "NullTime represents a time.Time that may be null.",
      "properties": {
        "Time": {
          "type": "string",
          "format": "date-time"
        },
        "Valid": {
          "type": "boolean"
        }
      },
      "x-go-package": "database/sql"
    },
...
    "UUID": {
      "description": "A UUID is a 128 bit (16 byte) Universal Unique IDentifier as defined in RFC\n4122.",
      "type": "array",
      "items": {
        "type": "integer",
        "format": "uint8"
      },
      "x-go-package": "github.com/google/uuid"
    }
  },
  "responses": {
    "GenericResError": {
      "description": "Unexpected error occurred",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "enum": ["error"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "GenericResFailConflict": {
      "description": "Server understand request but refuse to authorize it",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
    "GenericResFailNotFound": {
      "description": "Fail due to resource not exists",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseAdminUploadedQuiz": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "x-go-name": "CreatedAt"
                },
                "creator_id": {
                  "type": "string",
                  "x-go-name": "CreatorID"
                },
                "description": {
                  "$ref": "#/definitions/NullString"
                },
                "id": {
                  "$ref": "#/definitions/UUID"
                },
                "title": {
                  "type": "string",
                  "x-go-name": "Title"
                },
                "total_questions": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "TotalQuestions"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "x-go-name": "UpdatedAt"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseGenerateDemoSession": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "string",
            "x-go-name": "QuizId"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseGetQuizAnalysis": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "avg_response_time": {
                  "type": "number",
                  "format": "float",
                  "x-go-name": "AvgResponseTime"
                },
                "correct_answer": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "x-go-name": "CorrectAnswers"
                },
                "duration": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "DurationInSeconds"
                },
                "options": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "x-go-name": "Options"
                },
                "options_media": {
                  "type": "string",
                  "x-go-name": "OptionsMedia"
                },
                "question": {
                  "type": "string",
                  "x-go-name": "Question"
                },
                "question_id": {
                  "$ref": "#/definitions/UUID"
                },
                "question_media": {
                  "type": "string",
                  "x-go-name": "QuestionsMedia"
                },
                "resource": {
                  "type": "string",
                  "x-go-name": "Resource"
                },
                "selected_answers": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  },
                  "x-go-name": "SelectedAnswers"
                },
                "type": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Type"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseOkWithMessage": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseQuizCreated": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "string",
            "x-go-name": "QuizId"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "RsponseListQuizzesAnalysis": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Count"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "activated_from": {
                  "$ref": "#/definitions/NullTime"
                },
                "activated_to": {
                  "$ref": "#/definitions/NullTime"
                },
                "correct_answers": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "CorrectAnswers"
                },
                "description": {
                  "$ref": "#/definitions/NullString"
                },
                "id": {
                  "$ref": "#/definitions/UUID"
                },
                "participants": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Participants"
                },
                "questions": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Questions"
                },
                "title": {
                  "type": "string",
                  "x-go-name": "Title"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    }
  }
}